LUN-2727 Remove unnecessary checks.
There are valid cases that do not pass this checks.
Example: when the first related model (placeholder container) does not have an admin. That model could just be a link.
In this case the permissions have to be created in that way. (considering the link model, not the model that has the admin)